requirements/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-cleanup==5.0.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==3.0.5
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-gis==0.15
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
numpy==1.19.1
pandas==1.1.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.1.0
